refactor: extract MASM `project_support` into `artifact/project_support.rs`
